style: isolate types into their own files
